refactor: use photobank barrel for DTO imports
