+++ frontend/packages/shared/src/cache/photosCache.ts
@@ -2,7 +2,7 @@ import Dexie, { type Table } from 'dexie';
 import { LRUCache } from 'lru-cache';
 
 import { isBrowser } from '../utils/isBrowser';
-import type { PhotoDto } from '../api/photobank/model';
+import type { PhotoDto } from '../api/photobank';
 
 export interface CachedPhoto {
   id: number;
